update BodyPlanTesting_nn2 to integrate the new encodings. Warning this experiment needs further update

+++ experiments/bodyplantesting_nn2/ManLog.cpp
@@ -16,7 +16,7 @@ void TestsLog::saveLog(EA::Ptr &ea)
     for(size_t ind = 0; ind < ea->get_population().size(); ind++){
         logFileStream << generation * ea->get_population().size() + ind << ",";
         std::vector<bool> manRes = std::dynamic_pointer_cast<CPPNIndividual>(ea->getIndividual(ind))->getManRes();
-        for(int j = 0; j < manRes.size(); j++){
+        for(size_t j = 0; j < manRes.size(); j++){
             logFileStream << manRes[j] << ",";
         }
         //logFileStream << ea->get_population()[ind]->getObjectives()[0] << ",";
@@ -36,7 +36,7 @@ void morphDescCartWHDLog::saveLog(EA::Ptr &ea)
         logFileStream << generation * ea->get_population().size() + ind << ",";
         Eigen::VectorXd morphDesc = std::dynamic_pointer_cast<CPPNIndividual>(ea->getIndividual(ind))->get_feature_desc().to_eigen_vector();
         logFileStream << morphDesc(0);
-        for(int j = 1; j < morphDesc.size(); j++){
+        for(size_t j = 1; j < morphDesc.size(); j++){
             logFileStream << "," << morphDesc(j);
         }
         logFileStream << std::endl;
@@ -55,7 +55,7 @@ void SkeletonMatrixLog::saveLog(EA::Ptr &ea)
             return;
         std::vector<std::vector<double>> matrix_4d = std::dynamic_pointer_cast<CPPNIndividual>(ea->getIndividual(i))->get_matrix_4d();
         logFileStream << "angle" << ",";
-        for(int i = 0; i < matrix_4d.at(0).size(); i++){
+        for(size_t i = 0; i < matrix_4d.at(0).size(); i++){
             if(matrix_4d.at(0).at(i) < 0.00001 && matrix_4d.at(0).at(i) > -0.00001)
                 logFileStream << 0.0 << ",";
             else
@@ -63,7 +63,7 @@ void SkeletonMatrixLog::saveLog(EA::Ptr &ea)
         }
         logFileStream << std::endl;
         logFileStream << "skeleton" << ",";
-        for(int i = 0; i < matrix_4d.at(1).size(); i++){
+        for(size_t i = 0; i < matrix_4d.at(1).size(); i++){
             if(matrix_4d.at(1).at(i) < 0.00001 && matrix_4d.at(1).at(i) > -0.00001)
                 logFileStream << 0.0 << ",";
             else
@@ -71,7 +71,7 @@ void SkeletonMatrixLog::saveLog(EA::Ptr &ea)
         }
         logFileStream << std::endl;
         logFileStream << "wheel" << ",";
-        for(int i = 0; i < matrix_4d.at(2).size(); i++){
+        for(size_t i = 0; i < matrix_4d.at(2).size(); i++){
             if(matrix_4d.at(2).at(i) < 0.00001 && matrix_4d.at(2).at(i) > -0.00001)
                 logFileStream << 0.0 << ",";
             else
@@ -79,7 +79,7 @@ void SkeletonMatrixLog::saveLog(EA::Ptr &ea)
         }
         logFileStream << std::endl;
         logFileStream << "sensor" << ",";
-        for(int i = 0; i < matrix_4d.at(3).size(); i++){
+        for(size_t i = 0; i < matrix_4d.at(3).size(); i++){
             if(matrix_4d.at(3).at(i) < 0.00001 && matrix_4d.at(3).at(i) > -0.00001)
                 logFileStream << 0.0 << ",";
             else
@@ -87,7 +87,7 @@ void SkeletonMatrixLog::saveLog(EA::Ptr &ea)
         }
         logFileStream << std::endl;
         logFileStream << "joint" << ",";
-        for(int i = 0; i < matrix_4d.at(4).size(); i++){
+        for(size_t i = 0; i < matrix_4d.at(4).size(); i++){
             if(matrix_4d.at(4).at(i) < 0.00001 && matrix_4d.at(4).at(i) > -0.00001)
                 logFileStream << 0.0 << ",";
             else
@@ -95,7 +95,7 @@ void SkeletonMatrixLog::saveLog(EA::Ptr &ea)
         }
         logFileStream << std::endl;
         logFileStream << "caster" << ",";
-        for(int i = 0; i < matrix_4d.at(5).size(); i++){
+        for(size_t i = 0; i < matrix_4d.at(5).size(); i++){
             if(matrix_4d.at(5).at(i) < 0.00001 && matrix_4d.at(5).at(i) > -0.00001)
                 logFileStream << 0.0 << ",";
             else
@@ -115,7 +115,7 @@ void OrganPositionDescLog::saveLog(EA::Ptr &ea)
     for(size_t ind = 0; ind < ea->get_population().size(); ind++){
         logFileStream << generation * ea->get_population().size() + ind << ",";
         Eigen::VectorXd morphDesc = std::dynamic_pointer_cast<CPPNIndividual>(ea->getIndividual(ind))->get_matrix_descriptor().to_eigen_vector();
-        for(int j = 0; j < morphDesc.size(); j++){
+        for(size_t j = 0; j < morphDesc.size(); j++){
             logFileStream << morphDesc(j) << ",";
         }
         logFileStream << std::endl;
